# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePortfolio API",
        "description": "API para el sistema de portfolio electrónico de estudiantes",
        "contact": {
            "name": "ePortfolio Support",
            "email": "admin@eportfolio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "http://eportfolio.test/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://eportfolio.cifpcarlos3.es/api/v1",
            "description": "Production server"
        },
        {
            "url": "https://eportfolio.juezlti.eu/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/users/{parent_id}/evaluaciones-evidencias": {
            "get": {
                "tags": [
                    "EvaluacionEvidencia"
                ],
                "summary": "List all evaluacionevidencias",
                "description": "Retrieve a paginated list of evaluacionevidencias",
                "operationId": "d22fb9d05b198b43acb376fd96c87d2b",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EvaluacionEvidencia"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "EvaluacionEvidencia"
                ],
                "summary": "Create a new evaluacionevidencia",
                "description": "Create a new evaluacionevidencia resource",
                "operationId": "f308afa34023341a9f2ae4e5d3a32d36",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEvaluacionEvidenciaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EvaluacionEvidencia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{parent_id}/evaluaciones-evidencias/{id}": {
            "get": {
                "tags": [
                    "EvaluacionEvidencia"
                ],
                "summary": "Show a specific evaluacionevidencia",
                "description": "Retrieve a specific evaluacionevidencia by ID",
                "operationId": "999e63ae2c044145ac7145aa894aca0d",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the evaluacionevidencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EvaluacionEvidencia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "EvaluacionEvidencia"
                ],
                "summary": "Update a specific evaluacionevidencia",
                "description": "Update a specific evaluacionevidencia by ID",
                "operationId": "1938a5dc293f8dab150a122112a5eb28",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the evaluacionevidencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEvaluacionEvidenciaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EvaluacionEvidencia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "EvaluacionEvidencia"
                ],
                "summary": "Delete a specific evaluacionevidencia",
                "description": "Delete a specific evaluacionevidencia by ID",
                "operationId": "3a7988699ea666905942635270f71b2e",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the evidencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "EvaluacionEvidencia eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEvaluacionEvidenciaRequest": {
                "title": "Store Evaluación Evidencia Request",
                "description": "Datos requeridos para crear una Evaluación de Evidencia",
                "required": [
                    "puntuacion",
                    "estado"
                ],
                "properties": {
                    "evidencia_id": {
                        "description": "ID de la evidencia",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID del usuario que comenta",
                        "type": "integer"
                    },
                    "contenido": {
                        "description": "Contenido del comentario",
                        "type": "string"
                    },
                    "tipo": {
                        "description": "Tipo de comentario",
                        "type": "string",
                        "enum": [
                            "feedback",
                            "pregunta",
                            "sugerencia"
                        ]
                    }
                },
                "type": "object"
            },
...
            "UpdateEvaluacionEvidenciaRequest": {
                "title": "Update Evaluación Evidencia Request",
                "description": "Datos para actualizar una Evaluación de Evidencia",
                "properties": {
                    "evidencia_id": {
                        "description": "ID de la evidencia",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID del evaluador",
                        "type": "integer"
                    },
                    "puntuacion": {
                        "description": "Puntuación otorgada (0-10)",
                        "type": "number",
                        "format": "float",
                        "maximum": 10,
                        "minimum": 0
                    },
                    "estado": {
                        "description": "Estado de la evaluación",
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "completada",
                            "revisada"
                        ]
                    },
                    "observaciones": {
                        "description": "Observaciones de la evaluación",
                        "type": "string"
                    },
                    "fecha_evaluacion": {
                        "description": "Fecha de evaluación",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "EvaluacionEvidencia": {
                "title": "Evaluación de Evidencia",
                "description": "Modelo de Evaluación de Evidencia",
                "properties": {
                    "id": {
                        "description": "ID único",
                        "type": "integer"
                    },
                    "evidencia_id": {
                        "description": "ID de la evidencia",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID del evaluador",
                        "type": "integer"
                    },
                    "puntuacion": {
                        "description": "Puntuación otorgada",
                        "type": "number",
                        "format": "float"
                    },
                    "estado": {
                        "description": "Estado de la evaluación",
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "completada",
                            "revisada"
                        ]
                    },
                    "observaciones": {
                        "description": "Observaciones de la evaluación",
                        "type": "string"
                    },
                    "fecha_evaluacion": {
                        "description": "Fecha de evaluación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}